src/tycho/pyproject.toml
[project]
name = "tycho"
version = "0.1.0"
description = "The CSPLab Monolith"
readme = "README.md"
requires-python = "~=3.12.0"
dependencies = [
    "dependency-injector>=4.48.2",
    "django==5.2.7",
    "django-environ==0.12.0",
    "django-stubs[compatible-mypy]==5.2.7",
    "djangorestframework>=3.16.1",
    "gunicorn==23.0.0",
    "psycopg[binary,pool]==3.2.12",
    "pydantic>=2.12.4",
    "requests>=2.32.5",
    "sentry-sdk[django]==2.43.0",
    "types-requests>=2.32.4.20250913",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
  "django-debug-toolbar==6.1.0",
  "honcho==2.0.0",
  "mypy==1.18.2",
  "pytest==8.4.2",
  "pytest-cov==7.0.0",
  "pytest-django==4.11.1",
  "ruff==0.14.3",
  "types-requests>=2.32.4.20250913",
]

[tool.uv]
package = true

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"

[tool.uv.sources]
tycho = { path = ".", editable = true }

[tool.hatch.build.targets.wheel]
packages = [
  "apps",
  "tycho",
]

# Third party packages configuration
[tool.coverage.run]
omit = [
  "migrations/*",
  "tycho/asgi.py",
  "tycho/wsgi.py",
  "tycho/tests/*",
  "apps/**/tests/*",
]

[tool.pytest.ini_options]
addopts = "-v --cov-report term-missing --cov=apps --cov=tycho"
DJANGO_SETTINGS_MODULE = "tycho.settings"
python_files = [
    "test_*.py",
    "tests.py",
]
testpaths = [
    "tests",
    "tycho/tests",
    "apps/**/tests",
]

[tool.ruff]
target-version = "py310"
# Exclude a variety of commonly ignored directories.
exclude = [
  "migrations",
  ".venv",
]

[tool.ruff.lint]
select = [
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "D",  # pydocstyle
    "E",  # pycodestyle error
    "F",  # Pyflakes
    "I",  # Isort
    "ISC",  # flake8-implicit-str-concat
    "PLC",  # Pylint Convention
    "PLE",  # Pylint Error
    "PLR",  # Pylint Refactor
    "PLW",  # Pylint Warning
    "RUF100",  # Ruff unused-noqa
    "S",  # flake8-bandit
    "T20", # flake8-print
    "W",  # pycodestyle warning
]

[tool.ruff.lint.per-file-ignores]
"**/tests/*" = [
    "S101", # use of assert
]

[tool.ruff.lint.pydocstyle]
# Use Google-style docstrings.
convention = "google"

[tool.mypy]
ignore_missing_imports = true
plugins = ["mypy_django_plugin.main"]
exclude = [
  "migrations"
]
mypy_path = "."
namespace_packages = true
explicit_package_bases = true

[[tool.mypy.overrides]]
module = [
    "environ"
]
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "tycho.settings"
ignore_missing_model_attributes = true
